fixed a bug with frequency being rounded down in jobmanager.hpp; made single precision default; updated binary;

+++ radio/jobmanager.hpp
@@ -91,7 +91,7 @@ public:
         memcpy(m_buffer, buffer, length * sizeof(C));
     }
 
-    inline size_t frequency()const
+    inline double frequency()const
     {
         return m_frequency;
     }
